+++ src/main/java/jpj/boot/controller/GoodsController.java
@@ -78,7 +78,7 @@ public class GoodsController {
     }
 
     @RequestMapping("/edit")
-    public ModelAndView edit(Integer id) {
+    public ModelAndView edit(Long id) {
         ModelAndView mav = new ModelAndView("/goods/edit");
         Goods goods = goodsService.selectByPrimaryKey(id);
         if (goods == null) {
@@ -114,7 +114,7 @@ public class GoodsController {
 
     @ResponseBody
     @RequestMapping("/delete")
-    public boolean delete(HttpServletRequest request, Integer id) {
+    public boolean delete(HttpServletRequest request, Long id) {
         logger.info("userId:"+ HttpSessionUtil.getUserId(request.getSession())+"删除用户：id"+ id);
         return goodsService.deleteGoodsByPrimaryKey(id) > 0;
     }
