+++ src/main/java/jpj/boot/dao/GoodsMapper.java
@@ -5,13 +5,13 @@ import jpj.boot.entity.Goods;
 import java.util.List;
 
 public interface GoodsMapper {
-    int deleteByPrimaryKey(Integer id);
+    int deleteByPrimaryKey(Long id);
 
     int insert(Goods record);
 
     int insertSelective(Goods record);
 
-    Goods selectByPrimaryKey(Integer id);
+    Goods selectByPrimaryKey(Long id);
 
     int updateByPrimaryKeySelective(Goods record);
 
@@ -21,5 +21,5 @@ public interface GoodsMapper {
 
     Goods selectByCode(String code);
 
-    int deleteGoodsByPrimaryKey(Integer id);
+    int deleteGoodsByPrimaryKey(Long id);
 }
