+++ src/main/java/jpj/boot/dao/OutLibMapper.java
@@ -3,13 +3,13 @@ package jpj.boot.dao;
 import jpj.boot.entity.OutLib;
 
 public interface OutLibMapper {
-    int deleteByPrimaryKey(Integer id);
+    int deleteByPrimaryKey(Long id);
 
     int insert(OutLib record);
 
     int insertSelective(OutLib record);
 
-    OutLib selectByPrimaryKey(Integer id);
+    OutLib selectByPrimaryKey(Long id);
 
     int updateByPrimaryKeySelective(OutLib record);
