+++ src/main/java/jpj/boot/entity/TEnum.java
@@ -3,9 +3,9 @@ package jpj.boot.entity;
 import java.util.List;
 
 public class TEnum {
-    private Integer id;
+    private Long id;
 
-    private Integer pid;
+    private Long pid;
 
     private String code;
 
@@ -19,19 +19,19 @@ public class TEnum {
 
     private Integer desc;
 
-    public Integer getId() {
+    public Long getId() {
         return id;
     }
 
-    public void setId(Integer id) {
+    public void setId(Long id) {
         this.id = id;
     }
 
-    public Integer getPid() {
+    public Long getPid() {
         return pid;
     }
 
-    public void setPid(Integer pid) {
+    public void setPid(Long pid) {
         this.pid = pid;
     }
