+++ src/main/java/jpj/boot/entity/User.java
@@ -4,9 +4,9 @@ import java.io.Serializable;
 import java.util.Date;
 
 public class User implements Serializable {
-    private Integer id;
+    private Long id;
 
-    private Integer pid;
+    private Long pid;
 
     private String name;
 
@@ -24,19 +24,19 @@ public class User implements Serializable {
 
     private Date updateTime;
 
-    public Integer getId() {
+    public Long getId() {
         return id;
     }
 
-    public void setId(Integer id) {
+    public void setId(Long id) {
         this.id = id;
     }
 
-    public Integer getPid() {
+    public Long getPid() {
         return pid;
     }
 
-    public void setPid(Integer pid) {
+    public void setPid(Long pid) {
         this.pid = pid;
     }
