+++ src/main/java/jpj/boot/service/OrderService.java
@@ -8,13 +8,13 @@ import jpj.boot.entity.Order;
  */
 public interface OrderService {
 
-    int deleteByPrimaryKey(Integer id);
+    int deleteByPrimaryKey(Long id);
 
     int insert(Order record);
 
     int insertSelective(Order record);
 
-    Order selectByPrimaryKey(Integer id);
+    Order selectByPrimaryKey(Long id);
 
     int updateByPrimaryKeySelective(Order record);
